powerstackApi/admins/requirements.txt
requests~=2.31.0
boto3~=1.27.1
PyJWT~=2.8.0
pytz~=2023.3
botocore~=1.30.1